+++ examples/basic_dispatch.py
@@ -41,7 +41,7 @@ used at each step.
 
 from datetime import timedelta
 
-from odys import AssetPortfolio, EnergySystem, Generator, Load, LoadType, Scenario
+from odys import AssetPortfolio, EnergySystem, FixedLoad, Generator, Scenario
 from odys.results.optimization_results import OptimalDisptachResults
 from odys.utils.logging import get_logger, setup_rich_logging
 
@@ -61,7 +61,9 @@ def run_basic_dispatch() -> OptimalDisptachResults:
         nominal_power=150,
         variable_cost=0,
     )
-    load = Load(name="load", type=LoadType.Fixed)
+    load = FixedLoad(
+        name="load",
+    )
     portfolio = AssetPortfolio([ccgt, solar_pv, load])
 
     scenario = Scenario(
@@ -69,7 +71,7 @@ def run_basic_dispatch() -> OptimalDisptachResults:
             "ccgt": 24 * [100],
             "solar_pv": [0, 0, 0, 0, 0, 0, 10, 30, 60, 90, 110, 120, 125, 120, 110, 90, 60, 30, 10, 0, 0, 0, 0, 0],
         },
-        load_profiles={
+        fixed_load_profiles={
             "load": 24 * [70],
         },
     )