+++ tests/unit/test_optimization/constraints/test_generator_constraints.py
@@ -6,8 +6,8 @@ import pytest
 import xarray as xr
 from linopy.testing import assert_conequal
 
+from odys.domain.entities.fixed_load import FixedLoad
 from odys.domain.entities.generator import Generator
-from odys.domain.entities.load import Load
 from odys.domain.entities.portfolio import AssetPortfolio
 from odys.domain.scenarios import Scenario
 from odys.energy_system import EnergySystem
@@ -47,7 +47,7 @@ def generator2() -> Generator:
 def asset_portfolio_sample(
     generator1: Generator,
     generator2: Generator,
-    load1: Load,
+    load1: FixedLoad,
 ) -> AssetPortfolio:
     return AssetPortfolio(assets=[generator1, generator2, load1])
 
@@ -68,7 +68,7 @@ def energy_system_sample(
         timestep=timedelta(hours=1),
         scenarios=Scenario(
             available_capacity_profiles={},
-            load_profiles={"load1": demand_profile_sample},
+            fixed_load_profiles={"load1": demand_profile_sample},
         ),
     )